VORTEX-5414: Renaming heatmap package.
