+++ open-sphere-plugins/heatmap/src/main/java/io/opensphere/heatmap/DataRegistryHelper.java
@@ -1,4 +1,4 @@
-package io.opensphere.analysis.heatmap;
+package io.opensphere.heatmap;
 
 import java.awt.image.BufferedImage;
 import java.util.Collection;
