+++ open-sphere-plugins/heatmap/src/main/java/io/opensphere/heatmap/GeometryInfo.java
@@ -1,4 +1,4 @@
-package io.opensphere.analysis.heatmap;
+package io.opensphere.heatmap;
 
 import java.util.List;
