+++ open-sphere-plugins/heatmap/src/main/java/io/opensphere/heatmap/HeatmapMantleController.java
@@ -1,4 +1,4 @@
-package io.opensphere.analysis.heatmap;
+package io.opensphere.heatmap;
 
 import java.util.Collection;
 import java.util.List;
