+++ open-sphere-plugins/heatmap/src/main/java/io/opensphere/heatmap/HeatmapOptions.java
@@ -1,4 +1,4 @@
-package io.opensphere.analysis.heatmap;
+package io.opensphere.heatmap;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
