+++ open-sphere-plugins/heatmap/src/main/java/io/opensphere/heatmap/HeatmapVisualizationStyle.java
@@ -1,4 +1,4 @@
-package io.opensphere.analysis.heatmap;
+package io.opensphere.heatmap;
 
 import java.util.Arrays;
 import java.util.List;
