+++ open-sphere-plugins/heatmap/src/main/java/io/opensphere/heatmap/package-info.java
@@ -1,0 +1,2 @@
+/** Client-side heat map classes. */
+package io.opensphere.heatmap;
